+++ src/main/scala/org/ergoplatform/nodeView/wallet/scanning/Scan.scala
@@ -8,32 +8,7 @@ import scorex.util.serialization.{Reader, Writer}
 
 import scala.util.{Failure, Success, Try}
 
-object ScanWalletInteraction extends Enumeration {
-  type ScanWalletInteraction = Value
 
-  val Off = Value("off")
-  val Shared = Value("shared")
-  val Forced = Value("forced")
-
-  def toByte(v: Value): Byte = v match {
-    case Off => -1 : Byte
-    case Shared => -2 : Byte
-    case Forced => -3 : Byte
-  }
-
-  def fromByte(b: Byte): ScanWalletInteraction = b match {
-    case x: Byte if x == -1 => Off
-    case x: Byte if x == -2 => Shared
-    case x: Byte if x == -3 => Forced
-  }
-
-  def interactingWithWallet(v: Value): Boolean = v == Shared || v == Forced
-}
-
-object Tester extends App {
-  println(ScanWalletInteraction.withName("forced").id)
-  println(ScanWalletInteraction.withName("shared").id)
-}
 
 /**
   * Wraps information about user scan.
@@ -55,26 +30,6 @@ object Scan {
 
 }
 
-/**
-  * A class that encodes an API request to create a scan.
-  *
-  * @param scanName       - scan description (255 bytes in UTF-8 encoding max)
-  * @param trackingRule  - a predicate to scan the blockchain for specific scan-related boxes
-  *
-  */
-
-case class ScanRequest(scanName: String,
-                       trackingRule: ScanningPredicate,
-                       walletInteraction: Option[ScanWalletInteraction]) {
-  def toScan(scanId: ScanId): Try[Scan] = {
-    if (scanName.getBytes("UTF-8").length > Scan.MaxScanNameLength) {
-      Failure(new Exception(s"Too long scan name: $scanName"))
-    } else {
-      Success(Scan(scanId, scanName, trackingRule, walletInteraction.getOrElse(ScanWalletInteraction.Shared)))
-    }
-  }
-}
-
 object ScanSerializer extends ScorexSerializer[Scan] {
   override def serialize(app: Scan, w: Writer): Unit = {
     w.putShort(app.scanId)
@@ -94,6 +49,27 @@ object ScanSerializer extends ScorexSerializer[Scan] {
     }
     val sp = ScanningPredicateSerializer.parse(r)
     Scan(scanId, appName, sp, interactionFlag)
+  }
+}
+
+/**
+  * A class that encodes an API request to create a scan.
+  *
+  * @param scanName       - scan description (255 bytes in UTF-8 encoding max)
+  * @param trackingRule  - a predicate to scan the blockchain for specific scan-related boxes
+  * @param walletInteraction - how scan should interact with (p2pk) wallet, see @ScanWalletInteraction for details
+  *
+  */
+
+case class ScanRequest(scanName: String,
+                       trackingRule: ScanningPredicate,
+                       walletInteraction: Option[ScanWalletInteraction]) {
+  def toScan(scanId: ScanId): Try[Scan] = {
+    if (scanName.getBytes("UTF-8").length > Scan.MaxScanNameLength) {
+      Failure(new Exception(s"Too long scan name: $scanName"))
+    } else {
+      Success(Scan(scanId, scanName, trackingRule, walletInteraction.getOrElse(ScanWalletInteraction.Shared)))
+    }
   }
 }
